refactor header component
